+++ tradeexecutor/ethereum/lagoon/execution.py
@@ -60,6 +60,7 @@ class LagoonExecution(EthereumExecution):
             self.web3,
             strategy_universe,
             account_value_func=self.account_value_func,
+            web3config=self.web3config,
         )
 
         # TODO: Create a proper place fot his